feat(search): improve new search page + tests
