+++ src/features/search/components/SearchDetails.tsx
@@ -1,7 +1,7 @@
 import { t } from '@lingui/macro'
-import { useFocusEffect, useNavigation } from '@react-navigation/native'
-import React, { useCallback, useRef, useState } from 'react'
-import { NativeSyntheticEvent, TextInput, TextInputSubmitEditingEventData } from 'react-native'
+import { useNavigation } from '@react-navigation/native'
+import React, { useRef, useState } from 'react'
+import { TextInput } from 'react-native'
 import styled from 'styled-components/native'
 
 import { UseNavigationType } from 'features/navigation/RootNavigator'
@@ -21,46 +21,35 @@ import { useCustomSafeInsets } from 'ui/theme/useCustomSafeInsets'
 
 export const SearchDetails: React.FC = () => {
   const { top } = useCustomSafeInsets()
-  const [query, _setQuery] = useState<string>('')
-  const { searchState } = useSearch()
   const { navigate } = useNavigation<UseNavigationType>()
   const refSearchInput = useRef<TextInput | null>(null)
   const showResults = useShowResults()
-  const { searchState: stagedSearchState, dispatch: stagedDispatch } = useStagedSearch()
-
-  useFocusEffect(
-    useCallback(() => {
-      setQuery(searchState.query)
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [searchState.query])
-  )
-
-  function setQuery(value: string) {
-    stagedDispatch({ type: 'SET_QUERY', payload: value })
-    _setQuery(value)
-  }
+  const { searchState: stagedSearchState } = useStagedSearch()
+  const { searchState } = useSearch()
+  const [query, setQuery] = useState<string>(searchState.query || '')
 
   const resetSearchInput = () => {
     navigate(...getTabNavConfig('Search', { query: '' }))
     setQuery('')
   }
 
-  const onSubmitQuery = (event: NativeSyntheticEvent<TextInputSubmitEditingEventData>) => {
+  const onSubmitQuery = () => {
     // When we hit enter, we may have selected a category or a venue on the search landing page
     // these are the two potentially 'staged' filters that we want to commit to the global search state.
     // We also want to commit the price filter, as beneficiary users may have access to different offer
     // price range depending on their available credit.
     const { locationFilter, offerCategories, priceRange } = stagedSearchState
+
     navigate(
       ...getTabNavConfig('Search', {
         showResults: true,
-        query: event.nativeEvent.text,
+        query,
         locationFilter,
         offerCategories,
         priceRange,
       })
     )
-    analytics.logSearchQuery(event.nativeEvent.text)
+    analytics.logSearchQuery(query)
   }
 
   return (
